:green_heart: update test case and :tractor: code refactoring. utils.py is too lonely
