+++ test/test_constants.py
@@ -70,5 +70,7 @@ def test_map():
     value = [1]
     attr = ["渝水区"]
     map = Map("新余地图示例", width=1200, height=600)
-    map.add("", attr, value, maptype='新余市', is_visualmap=True, visual_text_color='#000')
-    print(map._repr_html_())
+    map.add("", attr, value, maptype='新余', is_visualmap=True,
+            visual_text_color='#000')
+    # To avoid potential pinyin crash, all cities have a province prefix
+    assert "jiang1_xi1_xin1_yu2" in map._repr_html_()
